# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/usuario/login":{"post":{"tags":["Login"],"summary":"Login","operationId":"login_api_usuario_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_usuario_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inventario/analisar-ia/{item_id}":{"post":{"tags":["Inventário"],"summary":"Analisar Item Com Ia","operationId":"analisar_item_com_ia_api_inventario_analisar_ia__item_id__post","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inventario/info-tecnica/{item_id}":{"get":{"tags":["Inventário"],"summary":"Obter Info Tecnica","operationId":"obter_info_tecnica_api_inventario_info_tecnica__item_id__get","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inventario/":{"get":{"tags":["Inventário"],"summary":"Listar Itens","operationId":"listar_itens_api_inventario__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InventarioRequest"},"type":"array","title":"Response Listar Itens Api Inventario  Get"}}}}}},"post":{"tags":["Inventário"],"summary":"Criar Item","operationId":"criar_item_api_inventario__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/inventario/{item_id}":{"get":{"tags":["Inventário"],"summary":"Buscar Item","operationId":"buscar_item_api_inventario__item_id__get","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Inventário"],"summary":"Atualizar Item","operationId":"atualizar_item_api_inventario__item_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventarioRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Inventário"],"summary":"Deletar Item","operationId":"deletar_item_api_inventario__item_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chatbot/chat":{"post":{"tags":["Chatbot"],"summary":"Chat","operationId":"chat_api_chatbot_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projetos/":{"get":{"tags":["Projetos"],"summary":"Listar Projetos","operationId":"listar_projetos_api_projetos__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjetoRequest"},"type":"array","title":"Response Listar Projetos Api Projetos  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Projetos"],"summary":"Criar Projeto","operationId":"criar_projeto_api_projetos__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/projetos/{projeto_id}":{"get":{"tags":["Projetos"],"summary":"Buscar Projeto","operationId":"buscar_projeto_api_projetos__projeto_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"projeto_id","in":"path","required":true,"schema":{"type":"string","title":"Projeto Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Projetos"],"summary":"Atualizar Projeto","operationId":"atualizar_projeto_api_projetos__projeto_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"projeto_id","in":"path","required":true,"schema":{"type":"string","title":"Projeto Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Projetos"],"summary":"Deletar Projeto","operationId":"deletar_projeto_api_projetos__projeto_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"projeto_id","in":"path","required":true,"schema":{"type":"string","title":"Projeto Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuario/newuser":{"post":{"tags":["Login"],"summary":"New User","operationId":"new_user_api_usuario_newuser_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/usuario/{email}":{"delete":{"tags":["Login"],"summary":"Deletar Usuario","operationId":"deletar_usuario_api_usuario__email__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuario/listaruser":{"get":{"tags":["Login"],"summary":"Listar","operationId":"Listar_api_usuario_listaruser_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_api_usuario_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_usuario_login_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InventarioInput":{"properties":{"Nome":{"type":"string","title":"Nome"},"Tipo":{"type":"string","title":"Tipo"},"Marca":{"type":"string","title":"Marca"},"Modelo":{"type":"string","title":"Modelo"},"Quantidade":{"type":"integer","title":"Quantidade"}},"type":"object","required":["Nome","Tipo","Marca","Modelo","Quantidade"],"title":"InventarioInput"},"InventarioRequest":{"properties":{"Id":{"type":"string","title":"Id"},"Nome":{"type":"string","title":"Nome"},"Tipo":{"type":"string","title":"Tipo"},"Marca":{"type":"string","title":"Marca"},"Modelo":{"type":"string","title":"Modelo"},"Quantidade":{"type":"integer","title":"Quantidade"},"Bateria":{"type":"string","title":"Bateria","default":""},"Coil":{"type":"string","title":"Coil","default":""},"Reservatorio":{"type":"string","title":"Reservatorio","default":""},"Sensores":{"type":"string","title":"Sensores","default":""},"Circuito":{"type":"string","title":"Circuito","default":""},"Outros":{"type":"string","title":"Outros","default":""}},"type":"object","required":["Id","Nome","Tipo","Marca","Modelo","Quantidade"],"title":"InventarioRequest"},"Message":{"properties":{"user_id":{"type":"string","title":"User Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["user_id","message"],"title":"Message"},"ProjetoRequest":{"properties":{"Id":{"type":"string","title":"Id"},"Nome":{"type":"string","title":"Nome"},"Resumo":{"type":"string","title":"Resumo"},"ComponentesNecessarios":{"type":"string","title":"Componentesnecessarios"},"NivelEscolaridade":{"type":"string","title":"Nivelescolaridade"},"Manual":{"type":"string","title":"Manual"},"Autor":{"type":"string","title":"Autor"}},"type":"object","required":["Id","Nome","Resumo","ComponentesNecessarios","NivelEscolaridade","Manual","Autor"],"title":"ProjetoRequest"},"UsuarioRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"senha":{"type":"string","title":"Senha"},"tipo":{"type":"string","title":"Tipo"}},"type":"object","required":["email","senha","tipo"],"title":"UsuarioRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/usuario/login"}}}}}}